pyproject.toml
[project]
name = "sql-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.3.2",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]
